Manual MAC, data types

+++ arduino-modbus-rtu-tcp-gateway/03-modbus-rtu.ino
@@ -24,13 +24,13 @@ void sendSerial() {
   if (queueHeaders.isEmpty()) {
     return;
   }
-  static int txNdx = 0;
+  static uint16_t txNdx = 0;
   header myHeader = queueHeaders.first();
   switch (serialState) {
     case 0:  // IDLE: Optimize queue (prioritize requests from responding slaves) and trigger sending via serial
       while (priorityReqInQueue && (queueHeaders.first().requestType & PRIORITY_REQUEST) == false) {
         // move requests to non responding slaves to the tail of the queue
-        for (byte i = 0; i < queueHeaders.first().msgLen; i++) {
+        for (uint8_t i = 0; i < queueHeaders.first().msgLen; i++) {
           queueData.push(queueData.shift());
         }
         queueHeaders.push(queueHeaders.shift());
@@ -80,7 +80,7 @@ void sendSerial() {
         myHeader.atts++;
         queueHeaders.shift();
         queueHeaders.unshift(myHeader);
-        unsigned long delay = localConfig.serialTimeout;
+        uint32_t delay = localConfig.serialTimeout;
         if (myHeader.requestType & SCAN_REQUEST) delay = SCAN_TIMEOUT;  // fixed timeout for scan requests
         sendMicroTimer.sleep(delay * 1000UL);
         serialState++;
@@ -93,17 +93,17 @@ void sendSerial() {
         } else if (myHeader.atts >= localConfig.serialAttempts) {
           // send modbus error 0x0B (Gateway Target Device Failed to Respond) - usually means that target device (address) is not present
           setSlaveStatus(queueData[0], SLAVE_ERROR_0B, true, false);
-          byte MBAP[] = { myHeader.tid[0],
+          uint8_t MBAP[] = { myHeader.tid[0],
                           myHeader.tid[1],
                           0x00,
                           0x00,
                           0x00,
                           0x03 };
-          byte PDU[5] = { queueData[0],
+          uint8_t PDU[5] = { queueData[0],
                           byte(queueData[1] + 0x80),
                           0x0B };
           crc = 0xFFFF;
-          for (byte i = 0; i < 3; i++) {
+          for (uint8_t i = 0; i < 3; i++) {
             calculateCRC(PDU[i]);
           }
           PDU[3] = lowByte(crc);  // send CRC, low byte first
@@ -123,10 +123,10 @@ void sendSerial() {
 }
 
 void recvSerial() {
-  static int rxNdx = 0;
-  static byte serialIn[MODBUS_SIZE];
+  static uint16_t rxNdx = 0;
+  static uint8_t serialIn[MODBUS_SIZE];
   while (mySerial.available() > 0) {
-    byte b = mySerial.read();
+    uint8_t b = mySerial.read();
     if (rxNdx < MODBUS_SIZE) {
       serialIn[rxNdx] = b;
       rxNdx++;
@@ -144,7 +144,7 @@ void recvSerial() {
       } else {
         setSlaveStatus(serialIn[0], SLAVE_OK, true, myHeader.requestType & SCAN_REQUEST);
       }
-      byte MBAP[] = {
+      uint8_t MBAP[] = {
         myHeader.tid[0],
         myHeader.tid[1],
         0x00,
@@ -164,7 +164,7 @@ void recvSerial() {
   }
 }
 
-void sendResponse(const byte MBAP[], const byte PDU[], const int pduLength) {
+void sendResponse(const uint8_t MBAP[], const uint8_t PDU[], const uint16_t pduLength) {
   header myHeader = queueHeaders.first();
   responseLen = 0;
   while (responseLen < pduLength) {  // include CRC
@@ -186,7 +186,7 @@ void sendResponse(const byte MBAP[], const byte PDU[], const int pduLength) {
     if (!localConfig.enableRtuOverTcp) ethCount[DATA_TX] += 4;
 #endif /* ENABLE_EXTRA_DIAG */
   } else if (myHeader.requestType & TCP_REQUEST) {
-    byte sock = myHeader.requestType & TCP_REQUEST_MASK;
+    uint8_t sock = myHeader.requestType & TCP_REQUEST_MASK;
     EthernetClient client = EthernetClient(sock);
     if (W5100.readSnSR(sock) == SnSR::ESTABLISHED && W5100.readSnDPORT(sock) == myHeader.remPort) {  // Check remote port should be enough or check also rem IP?
       if (localConfig.enableRtuOverTcp) client.write(PDU, pduLength);
@@ -203,9 +203,9 @@ void sendResponse(const byte MBAP[], const byte PDU[], const int pduLength) {
   deleteRequest();
 }
 
-bool checkCRC(byte buf[], int len) {
+bool checkCRC(uint8_t buf[], int16_t len) {
   crc = 0xFFFF;
-  for (byte i = 0; i < len - 2; i++) {
+  for (uint8_t i = 0; i < len - 2; i++) {
     calculateCRC(buf[i]);
   }
   if (highByte(crc) == buf[len - 1] && lowByte(crc) == buf[len - 2]) {
@@ -215,9 +215,9 @@ bool checkCRC(byte buf[], int len) {
   }
 }
 
-void calculateCRC(byte b) {
+void calculateCRC(uint8_t b) {
   crc ^= (uint16_t)b;              // XOR byte into least sig. byte of crc
-  for (byte i = 8; i != 0; i--) {  // Loop over each bit
+  for (uint8_t i = 8; i != 0; i--) {  // Loop over each bit
     if ((crc & 0x0001) != 0) {     // If the LSB is set
       crc >>= 1;                   // Shift right and XOR 0xA001
       crc ^= 0xA001;
